Set async for kubeadm upgrade node command
Although this command should timeout by itself, we often see that these
tasks run forever in the CI (until the hard limit of the CI job hits).
Hopefully, setting async on these tasks fixes that. Although the command
may still fail, at least we do not block the CI for so long in that case.