Ensure ca.key is absent on CP nodes
This fixes that these files may can't get removed by an un-privileged
Ansible user.

These keys may still lie around for clusters which have been created
before !553. Under normal circumstances this does not do any harm,
but during root CA rotations this can get in your way as kubeadm will
notice that this file is present and may sets unwanted configurations
then. More detailed, we noticed that the kube-controller-manager manifest
got flags pointing to that file on Kubernetes upgrades. This is bad if
the key does not fit the CA anymore because the root CA already got rotated.
As we manage the CA inside of an Hashicorp Vault instance, we
want to ensure it is absent in any case and lies not around orphaned.