Add hotfixing strategy
To be able to provide hotfixes for the current as well as
for older releases we need to define a strategy to do so.
For this we need to create some additional ci-jobs and
alter the pipeline.

As we again use towncrier for the changelog-creation we need to put the
releasenote in a seperate directory and use a different config (which
will be generated within the ci) to not genereate the changelog from all
current notes present but only for the hotfix.