Extend CI cleanup script for volume snapshots
As we're testing the snapshot-controller in our smoke tests, it can
happen that a volume snapshot is left over if these tests fail.
However, the corresponding volume then cannot automatically be deleted
if a snapshot is still around. Therefore, we're extending the script
such that volume snapshots get deleted before volumes.