Restrict rules for containerd job
Run this job only on merge-trains (if test-rules-changes apply) and on
release-prepare-branches. Otherwise create a manual job. This prevents
it from getting executed for each push on the "devel" branch (among
others) which is reasonable as everything should go through a merge
train once and saves us from unnecessarily waiting for this job to
finish until the pages job is executed (and docs get build). We
introduced release management and devel now really is the development
branch, therefore we must not ensure that the containerd job is executed
on devel again.