Ignore failures for GPU tests
Unfortunately, there is no easy way to figure out if a node has a free
GPU or not. The allocatable status does not reflect the actual
utilization. Therefore, we're disarming the tests by reducing the
wait timeout and ignore failures as the Pod will get stuck in Pending if
no GPU is available on the node. For fixing the root cause, #610 got
re-opened.