Drop manifest-based Calico installation
Support for the manifest-based Calico installation is getting dropped as
we're enforcing the usage of the Tigera operator since Kubernetes v1.26.
As we dropped support for K8s v1.25 in !1056 , we're now at a point in
time where we are able to completely drop it. The manifest-based
installation procedure was deeply nested into the LCM. Documentation has
been adjusted, installation procedure has been adjusted, the
root-ca-rotation feature has been updated as well as smoke tests.

As we managed the certificates of Calico components via Vault for the
manifest-based installation but not for the operator-based installation
(yet), the Calico PKI engine has been dropped from Vault. A check has
been answered which warns if the engine is still around and has a valid
CA.

The vault policies have been marked as deprecated and should be dropped
in the near future (rel !1076).

This step has been announced in:
https://gitlab.com/yaook/k8s/-/blob/c1042793a2a9541fb9de31cfdc18ef97717de829/CHANGELOG.md?plain=1#L7-16