rook_v2: Allow to unset CPU limits
Previously, it was not possible to unset CPU limits for these as even if
not defined, the helm chart default values would have been used. So we
must explicitly overwrite the value with "null".

To pass null without brackets, we have to use the to_json filter when
supplying these values. Therefore this is changed everywhere.

Not done for rook_v1 as its deprecated and will be dropped in !1042.

Closes: #627