flux: Fix PodMonitor creation
The "monitoring_install" variable may not be around. In contrast, the
"k8s_monitoring_enabled" variable must be set and is therefore always
around. This fixes the issue that despite monitoring being disabled,
the flux chart couldn't get deployed as it tried to create PodMonitors
which is not possible in that case as Prometheus CRDs are not around.