Fix nuking an OpenStack cluster with Gitlab as backend
When nuking a cluster, also object storage buckets are attempted to be
removed. However, the previous code tried to access the local Terraform
state file which is not around when Gitlab is used as Terraform backend.
The correct way is to process the remote Terraform state in that case.
This has been overseen in !998.