Fix kubernetes-validate version during K8s upgrades
We restricted the kubernetes-validate version on K8s nodes in !1004 for
good reasons. However, during a Kubernetes upgrade it now happens that
as long as the Kubernetes version in the config/config.toml is still the
previous Kubernetes version, the kubernetes-validate package on the nodes
does not get updated. This can lead to issues if an upgrade has to be
aborted and restarted. This commit configures the kubernetes-validate
package such that the next version is installed during Kubernetes
upgrades.
On the long run we want to address #617

Closes: #632