Add check whether VAULT_TOKEN is set
Stages 2 and 3 need a vault token with at least orchestrator role. It is
easy to forget setting the variable, so the run will stop some minutes
into the run. Time that is better spent otherwise. This commit adds a
check to both stages so they fail if the variable is not set.