Add policy for release support
As we have multiple older releases now we need to define how long we
will support them with hotfixes. As discussed in
https://gitlab.com/yaook/k8s/-/issues/633 we define a number of old
releases and a minimum time-period for support.

As the suggestion was to support for at least two weeks after EOLA but
we decided on not using EOLA I defined it to be four weeks after
release (which is an increase by the current medium time between two
releases).