Install CCM and cinder csi plugin via helm
Previusly we installed the openstack csi-components via manifests. We
switch to helm charts now to enable easier updates. The migration can be
enforced when adding support for future Kubernetes versions.

To allow old clusters to switch to the helm based installation of CCM and
cinder csi plugin we introduce a migration step to delete the old
ressources and reinstall them with helm.

To allow an OpenStack credential rotation (which requires triggering
handlers if the cloud-config secret changes), we do manage the
cloud-config secret on our.