Thanos: Allow to set in-memory sizes
When configuring a memory limit on the storegateway component the Pod is
getting OOM killed on high load as Thanos is not aware of that limit.
However, Thanos allows to configure its so called in-memory cache [1] which
should prevent the storegateway component from exceeding its Kubernetes
wise memory limits on high load.

The in-memory cache can also be configured via a dedicated helm chart
value [3]. However, that leads to a crashlooping storegateway as the
ConfigMap does not get created and is not passed as volume to the
StatefulSet by the helm chart [4] (trigger warning, stale closed issue).
Therefore, we're choosing the extraFlags.

It would make sense to choose the values to be supplied based on the
existing Kubernetes memory limits (thanos_store_memory_limit variable).
However, Thanos does not accept binary units of measurement (e.g. '2Gi')
and it is not doable to convert them easily with Ansible as the
'human_to_bytes' filter [5] does not support binary units as well.
Therefore, additional variables have been created and it is in the
operators responsibility to define sane values with fitting units.

Closes: #635

[1] https://thanos.io/tip/components/store.md/#in-memory-index-cache
[2] https://github.com/bitnami/charts/blob/b0633a84f4ce0cb9dfe907272f48373b87e13b06/bitnami/thanos/templates/storegateway/statefulset.yaml#L117-127
[3] https://github.com/bitnami/charts/blob/b0633a84f4ce0cb9dfe907272f48373b87e13b06/bitnami/thanos/values.yaml#L88
[4] https://github.com/bitnami/charts/issues/14362
[5] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/human_to_bytes_filter.html