Init yk8s with latest release by default
This change introduces a new env var `$MANAGED_K8S_LATEST_RELEASE`
for the `init.sh` script which is `true` by default and causes
the init script to initialize the managed-k8s submodule with the
branch of the latest version.
The latest version is determined by the `version` file in the
cloned copy of yaook/k8s that is used to run the `init.sh` script.

To not affect the CI, the variable is set to `false` in the CI
configuration.

Closes #625