Change ci rules
We used $CI_COMMIT_REF_NAME == devel in our rules. As described in [1]
this means we allow a pipeline for every branch or tag called devel. As
we only want to run this for the branch we change to CI_COMMIT_BRANCH.

It's likely we never will have a tag called "devel", but this is a change
for cleaness, consistency and easier readability.

[1] https://docs.gitlab.com/ee/ci/variables/predefined_variables.html