Clean up ipsec installation
Too many packages were installed. It's sufficient to just install
"charon-systemd" and its dependencies.

With the previous list of packages, the legacy setup procedure has been
installed as well. That led to the problem, that the connection can't be
initialized after reboots as the ipsec service (alias strongswan-starter
which is the legacy procedure) was around and enabled and interfered
with the strongswan service (modern procedure) interfere and both ran in
rivalry on reboots.