Make storageClass creation for pools optional again
The attribute "create_storage_class" of a pool has been introduced with
rook_v1 [1] and is an optional attribute. However, when implementing the
helmification of rook/Ceph, the creation of storage classes for ceph
pools has been hard coded. There are real life use cases where that's
not desired, e.g. if the Ceph cluster is used by an external OpenStack
deployment. Although that variable was previously a string, it always
has been interpreted as boolean. Therefore, I changed that directly.

No changes for the rook_v1 role is legacy and will get removed in
!1042 anyway.

[1] https://gitlab.com/yaook/k8s/-/blob/f93f8bda18d3cd4ff12d1201725781f2632ac589/k8s-service-layer/roles/rook_v1/tasks/main.yaml#L186