Use general on_openstack variable for rook
Previously, with the old repository layout, we did not pass the hosts
file when invoking the k8s-service-layer playbooks and therefore
variables defined there were not accessible by e.g. the rook roles.
With the new layout that's not the case anymore and we can make use of
these and therefore remove the "rook_on_openstack" variable as that was
just a workaround to tell rook if we're running on OpenStack or not.

For clusters running on OpenStack, the "on_openstack" variable is set by
Terraform in the hosts file. For clusters not running on OpenStack, it
must be set to false in the hosts file anyway.