Fix Prometheus stack values templating
The affinity and tolerations had the wrong indentation. This one the
one side led to the problem that a configured "scheduling_key" did not
take effect for the Prometheus Pod and that the deployment failed at all
if additionally "allow_external_rules" has been configured as the
rendered template was not a valid yaml file.