Streamline Thanos bucket management
It's confusing that the default value on the Terraform level is false,
although we're patching it to true via the update_inventory.py script.
The reason for that is kind of historical and to circumvent that a
bucket for Thanos gets's created although Thanos itself is disabled.
That's something we need to handle on a higher level than Terraform
itself.

However, documentation was also kind of misleading. This commit should
make things a bit more clearer.

One downside of moving the Thanos bucket configuration to vault is, that
we can't easily check if a bucket is configured or managed and fail if
none applies in the update_inventory.py script.