Prevent Terraform disruption through the config
Adds the `prevent_disruption` setting in the `[terraform]` config section.
 When set to `true`, which is the default, Terraform is stopped before
 performing disruptive actions.
When `TF_USAGE=true` this also overrides the `MANAGED_K8S_DISRUPT_THE_HARBOUR`
 environment variable.

This is a breaking change because `[terraform].prevent_disruption` has to be
 actively disarmed in order to receive the previous behavior.

Adds `yq` as dependency to read the config.toml in bash.

Breaking: true
Part-of: #647