Use short-lived kubeconfigs on orchestrator
Currently, the kubeconfigs of the orchestrator are pulled from /etc/
kubernetes/admin.conf on a control plane node. This file is valid for 1
year and is regularly encrypted and checked into version control as part
of the cluster repo.

This commit changes the behavior of the k8s-master role to use the
pre-existing script k8s-login.sh to create a short lived kubeconfig with
admin privileges from the Vault. It also adds the path of the kubeconfig
to .gitignore.