vault: autodetect cluster name
Previously, most scripts in tools/vault expected the cluster name as
their first argument. The cluster name is configured in config.toml
at vault.cluster_name from where it had to be manually read to run
the scripts. This commit leverages Python to enable the scripts to
read the cluster name from the config directly. An additional check
is added that lets the script fail if the cluster name is not set and
asks for confirmation if the cluster name is set to the default value
"devcluster".