Use direnv layouts in yk8s repo
Right now, the direnv layouts for Nix Flakes and Python Poetry are only
loaded from the cluster repository. This means that a simple checkout
of yaook/k8s without cluster repo (eg. for development purposes) may
not have access to all its dependencies. This commit adds both direnv
layouts to the .envrc of the yaook/k8s repository itself.

In order to prevent the layouts from being triggered multiple times,
guards are added to each of them so they are skipped if they are already
loaded.