Build pages only on devel
As sphinx-multiversion uses the config from the current branch to build
the docs for all releases, we need to restrict the job to build only on
devel and no longer on the release branches. That is because we added
new extensions which are needed to build the latest version. Thise
extensions aren't present in the older configs.