Switch prepare gateway order
Otherwise, if ipsec is disabled, the handlers for bird and nftables are
skipped which i.e. prevents the gateways from exchanging routes with the
K8s control plane nodes. bird and nftables are both dependencies of the
ipsec role. It is unclear why the handlers are skipped (a notification
is saved correctly when the bird and nftables role are executed) just
because the ipsec role is skipped. This seems to be new Ansible behavior.