Talk to k8s API via orchestrator only
Some tasks have been triggered from the first control plane node. This
has downsides. E.g. this node can be down connectivity wise. Also to be
able to trigger Ansible tasks with modules from the Kubernetes.Core
collection we had to deploy some pip packages on the control plane nodes.
This has been done by the k8s-tools role which is now obsolete.
There is an urging need to switch now, as newer versions of the
kubernetes-validate package (which was deployed on the control plane
nodes for that reason) have misconfigured metadata and can not be
installed. Related issues already have been addressed in !1097 & !1004.
Instead of implementing another workaround, this commit has been packed.
This frees us from the burden to maintain pip dependencies on the cp nodes.