Drop support for rook_v1
... and simplify version checks as they were a bit too complex.
We can simply drop rook_v1 as it itself and the rook version it was able
to deploy are deprecated a long time.
The variable for disruption has been missing, it is needed to perform
rook/ceph upgrades.
The manage_pod_budgets variable had no default. It was previously loaded
from the rook_v1 role.