WIP: tf: Provide control over the whole node name
Allows the user to configure the whole name of nodes.
Controller and worker nodes are now stored together in the `terraform.nodes`
 parent table and distinguished by the `role` attribute [1].
The full node name is given as the table name while for gateway nodes a
 common name can be configured.
 If a cluster name is configured it is prefixed to all node names.

This is a non breaking change in terms of behavior.

[1] Caveat: Changing the role of a Terraform node will rebuild the node
    which might be unexpected.

TODO: Provide migration script
      Change tables [terraform.controllers|workers.*] to
       [terraform.nodes.controller|worker-*] + role="controller|worker"

Breaking: true