Refine PKI renewal during K8s upgrades
Before triggering the actual Kubernetes upgrade, the certificates,
kubeconfigs et al. are verified and renewed on all Kubernetes nodes if
necessary. This step can now be explicitly triggered or skipped via the
renew-pki tag. This can save some time if an upgrade has to be aborted
and restarted.
Renewal depends on the helpers/vault-approle role which tasks therefore
are tagged with always to keep things simple.