Remove service-account-issuer patch for kube-apiserver
Introduced in !749, this patch was necessary to migrate the
service-account-issuer of existing clusters to an OIDC conformant HTTPS
URL in a flawless manner and make the Kubernetes distribution pass
sonobuoy conformance tests.
We can safely assume that Kubernetes clusters >= v1.26 use the patched
format. As we support no versions below, we can simply drop these tasks.