Remove option for not providing MRIID in releasenote
As the option of not providing a MR-IID often leads to errors in the
automated creation of the changelog (and as we don't see a reason for
not providing the information) we remove the option of setting '+' for
not replacing the filename.

Because of that change we removed the automated creation of a
releasenote when publishing a release.

Adresses: #642