Move all cluster configuration to config.toml
Most of the cluster configuration resides in config.toml but some things
are configured via environment variables instead (in most cases using
.envrc) which is not very obvious for newcomers.

The reason for this is that historically, all variables that needed to
be present in bash scripts were configured via the environment. However,
as we've recently introduces tomlq in order to read from the config.toml
in bash scripts, this restriction is overcome.

Thus, this commit reworks some variables to be configured inside
config.toml instead of .envrc, working towards the goal of having the
whole information about _how the cluster looks_ inside config.toml and
only the bits about _how the user interacts with the cluster_ in the
environment variables.