Add keyring backend variable to env template
There is a bug in poetry that makes it access the user's key ring even
if not necessary. It should only do that (1) when publishing or (2) when
using dependencies from private repositories. This can lead to problems
when the key ring is not available or is password protected.

This commit adds a variable setting to the template of
~/.config/yaook-k8s/env that avoids that problem.