Don't gitignore inventory
This commit modifies template.gitignore to keep the cluster repo inventory.
Most but not all variables in the inventory are generated from the
config.toml vie the update_inventory.py action script. However, for
clusters running on OpenStack, there are also files generated by
Terraform in the inventory. If these are not checked in into git, one
must execute the terraform stage in advance to other actions. Checking
in the whole inventory prevents that.

Closes: #664