Add LBaaS configuration for OVN compatibility
The fixed IP of the VIP port and the port IDs of the gateway nodes need
to be stored inside the controller-config.toml of the ch-k8s-lbaas so
the VIP port IP and the fixed IP addresses of L3 ports can be added to
the allowed address pairs of each gateway node port by the ch-k8s-lbaas.

In addition, it is important to also allow "0.0.0.0/0" if IPv4 is
enabled and "::/0" if IPv6 is enabled, because otherwise the agents
would drop these from the OpenStack ports which would restrict traffic
and leads to connectivity issues.