poetry: disable package mode
By default, Poetry tries to install the package defined in
pyproject.toml in addition to its dependencies. However, we use it only
for the dependencies and yk8s is only a dummy project. In order to avoid
installing the non-existant root package, we added --no-root to the
poetry arguments wherever it is called, but recent versions support a
configuration option to disable it project-wide.

Now that Nix is a requirement and poetry is installed via Nix for
everybody, we know that this option is available for everybody, too.