Enable manual SSH host key signing and verification
Disabling SSH host key verification is insecure. To simplify host key
verification, we make use of the ssh secrets engine of vault. For each
public key of a node, we generate a signed certificate. A proper ssh
known hosts file is generated and put into the etc directory. We have to
make use of TOFU as we must be able to connect to freshly deployed
nodes. Host key verification has been enabled in the Ansible config.