Enable automatic SSH host key signing and verification
Clusters (in especially dev-clusters) using vault in docker, have to
manually renew SSH certificates as the nodes are not able to talk to the
Vault instance.
For clusters where nodes can reach the vault instance, the renewal of
SSH certificates can be automated. For nodes to be able to talk to a
vault instance, the vault-cli tool has to be installed. The
vault-onboarded role has been adjusted such that it prepares and
deploys a vault-login script on each node. The deployed systemd service
and timer for the ssh cert renewal makes use of this script.

In order to be able to properly connect to the nodes, we have to pass
the list of keys and certs to devsec ssh hardening such that it can
create a proper sshd configuration. The role has been moved to the
connect-to-nodes playbook therefore.

We make use of resources already given by the metal-controller [1] with
some slight adjustments.

[1] https://gitlab.com/yaook/metal-controller/-/blob/349ea4ac1b574d6cdd15baa00ca75cd7ad212b69/metal_controller/configdrive/static/systemd/renew-ssh-certificates.timer