Speed things up
Skip connection roles and tasks if they already have been executed
during a run. As we're importing this playbook in advance of other
playbook, the connect-to-nodes playbook is executed quite often which
really slows down execution time. By setting cacheable facts, we can
skip roles and tasks if the playbook is just imported (cacheable is
needed to do so).
Also adjust the SSH flags such that connections are reused and persist.