Enhance Cinder CSI topology feature gate
We need to make sure that the Cinder CSI-Provisioner topology feature
gate can be toggled on and off. Since the feature is enabled by
default, we need to handle both states.

The Jinja template has been extended. This ensures compatibility with
enabled or disabled topology feature gate, defaulting to false.