Generate new kubeconfig before each root CA rotation step
We need to renew the kubeconfig before each root CA rotation step as we
now use short-lived kubeconfigs and do not pull the admin kubeconfig
from the CP nodes anymore. If not renewed, connection to the K8s API
gets lost in the second phase of the root CA rotation.