Adjust super admin kubeconfig handling
Add support to the k8s-login procedure to generate a super admin
kubeconfig. Such scoped kubeconfigs are necessary for kubeadm to be
able to initialize a cluster so we can't restrict the vault policies
for the control plane here and must ensure a super admin kubeconfig is
around when initializing the Kubernetes itself. Afterwards, we can
remove it from the nodes for security reasons.

Due to the modulartiy and the scopes of k8s-core and k8s-supplements, we
must ensure adjusting the scope of the kubeconfigs on the control plane
nodes is handled inside of k8s-core.