Fix CCM helm installation for fresh clusters
and set a priority class name and remove cloud-config related values.

The default for the dnsPolicy changed from "Default" to
"ClusterFirstWithHostNet" which prevents the CCM from coming up if the
coreDNS Pods are not running yet which is the case if the cluster got
freshly initialized.
https://github.com/kubernetes/cloud-provider-openstack/issues/2611

As we manually create the cloud-config secret, we should not
additionally set it via helm.

A priority class name prevents preemption.