Fix K8s upgrade CP targeting
The previous code didn't do what we expected but skipped all CP nodes
but the last due to no hosts matching. I am unsure why this was able to
pass the CI (i.e. the upgrade jobs).

I could not find a way to target all but the last node of a group so I
switched it and instead we're targeting every node but the first and
then the first.