Fix image tagging for releases
Release branches are named like this: "release/vX.Y.Z". This is not a
valid image tag (due to the /). This commit adjustes the deploy images
job such that they are split into one job for the default branch and one
for release branches. Both share a common template. For releases, we're
using the version contained in "./version" as image tag instead of the
target branch name now.