Add dispatch playbook for custom stage
Currently, the entrypoint for the custom stage lies outside of the LCM
but relies on the connect-to-nodes play. This results in issues whenever
something changes there requiring manual adaption in the custom
main.yaml.

This commit adds a new playbook that ensures connection to nodes and
then dispatches to the custom playbook outside of the LCM. Hence, the
changes required in the custom playbook can now be included in the
LCM changes themselves.