Fix Wireguard setup for dual stack
We don't enforce setting ip_cidr for a Wireguard endpoint if IPv4 is
enabled nor do we enforce ipv6_cidr to be set for a Wireguard endpoint
if IPv6 is enabled. We therefore have to adapt dependent code on if
these are set.
Depending on which IP family is enabled, we also must properly template
Wireguard configuration files.