Adjust and clean up bird configs for IPv6
The scoping was needed when connecting two clusters via a Wireguard
site-to-site tunnel. However, support for that has been removed. This
was a leftover.

For IPv6, we have to insure everything is imported from the kernel.
Otherwise, routes to Pod IPs announced via BGP get filtered as bird
is not aware that it actually can reach the Kubernetes worker via IPv6.
This is because the route to the internal IPv6 network is not static but
expires. It is therefore considered an "alien" route to bird and
filtered out by default. The changes prevent that.