Rework check-services role
Make names more clear.
Also we want to check services even though dual stack is not configured.
Therefore, checks are executed based upon ipv4, ipv6 or both are
enabled.
We should also check if the services have proper IP addresses before
actually trying to validate the connection.
As the image automagically configures itself and needs write access to
the filesystem to do so, we're now using the user 101 (nginx).

The smoke tests services are of type NodePort. We therefore want to
reach out to them via e.g. the first worker node.