Add check that at least one IP family is enabled
To do so, we must adjust the handling of ipv[4,6]_enabled variables. We
do not parse them into the inventory anymore but directly into the hosts
file. This is also stringent to bare metal clusters where the
dualstack_support variable had to be set in the hosts file.
Moving it to the hosts file allows us to check that at least one IP
family is enabled in all action scripts which trigger Ansible
independent of the undercloud (Terraform or bare metal).