Fix apiserver backend for modern haproxy versions
We used to hide a Host header behind the version string. Haproxy 2.2
introduced `http-check send` which is now the preferred way to add
headers to health checks. Haproxy 2.8 (which is in Ubuntu 24.04) now
throws an error when the header hack is used.

This commit replaces the hack by a `http-check send` instruction.

Starting with this commit the minium required haproxy version is haproxy
2.2 which means that the minimum Debian version is now Debian 11.