Add support for new ch-k8s-lbaas versions
which can handle DualStack-Services and IPv6 addresses to some kind of
extend, so also smoke tests have been adjusted.
This commit also makes the ch-k8s-lbaas version an optional
configuration option and created a default. Users then do not have to
manually adjust the version in their cluster configurations anymore but
can just unset it such that the default is used. This ensures users use
the latest version by default.