Do not wait for rook helm charts
Depending on the actual size of the cluster and other factos like
connectivity it can take a significant time for the Ceph cluster to come
up and get ready fully. If we do wait here, we may run into timeouts.
If the cluster is upgraded, there are upgrade checks triggered
afterwards. If the cluster is freshly ininitialized, we do have tasks
which check that the cluster is ready. So there is no need to wait here
and potentially run into timeouts.