Include k8s-config for K8s upgrade certificate renewal
Sadly, although config/k8s-config is a dependency of the k8s-master and
k8s-worker role, it is not included by "include_role". As variables like
the Pod and Service CIDRs must be known for certificate renewal, we have
to manually include it. Because "role:" takes precedence over "tasks:",
the variables will be available for the PKI renewal tasks.