Split off openstack from terraform section
The Terraform section was introduced with the assumption that yaook/
k8s will be run on OpenStack. This is not the case anymore and some
options can now either be set in terraform or miscellaneous.

In case of bare-metal, the Terraform section is not used, but some
networking specific options (like subnet_cidr) still need to be set.
Also, there might be cloud platforms other than OpenStack where
Terraform is used.

Thus, as discussed in #687:

> A new section openstack is introduced that contains all OpenStack
> specific options. More general options MUST NOT reside in the
> openstack section.