Use static facts instead of dynamically gathering facts
Instead of using helper roles which are triggered during and increase
the runtime, this commit adjusts code such that Ansible static facts are
used. If a node has a GPU attached, this fact can be gathered during
bootstrapping and it can not change afterwards (because adding a GPU to
a node which already joined Kubernetes does not work as that requires
preparations before the join). Once frontend services are initialized,
that fact should not change anymore. Same goes for the fact that a node
already joined Kubernetes. So there is no need to dynamically figure out
these facts on every run but we can make use of static facts instead.