Rework connection and bootstrap plays
Move roles which are not essential to establish node connectivity (so
basically everything besides detect-user) to the bootstrap playbook.
This has the advantage that these tasks and roles are not executed
everytime the connect-to-nodes playbook is included and may
significantly decrease runtime performance for install-all runs.
However, this has the downside that we potentially do stuff against
nodes which have not been bootstrapped, yet. This can for example
happen, if playbooks are directly triggered but install-all has not been
executed once (fully). Therefore, we're explicitly checking that every
node has been bootstrapped at least once with the help of a static fact.