Allow enabling announcement of service IPs
This is useful when you are dealing with a bare metal setup where
ch-k8s-lbaas cannot be used (for whichever reason). In that case, you
may not be able to use pod IPs in your frontend's reverse proxy and
using cluster IPs is more feasible.

However, calico does not announce the service IP range by default via
BGP: it needs to be explicitly configured.

This newly introduced toggle allows to enable that feature.