direnv: only load necessary dependencies by default
Currently, the flake devShell contains k8s dependencies plus some
interactive tools that have been added as a convenience. However, it is
possible that they cause problems on non NixOS systems (eg. ping not
having SUID).

This commit reduces the default devShell to only load the yk8s
dependencies and adds a second devShell "withInteractive" that comes
with the additional tools. It also adds a new environment variable
YAOOK_K8S_DEVSHELL that can be used to select which shell should be
loaded.

Also, the CI image is now built without the interactive dependencies, in
order to better detect cases in which required dependencies are
accidentally only added as optional.