docs: Add hint for dev vault setups
It is essential to set USE_VAULT_IN_DOCKER to true for development
setups as otherwise this condition [1] gets bogus and the nodes will
try to talk to a local vault to renew their SSH certificates.

Additionally some cosmetic changes have been made and the steps got
re-arranged a bit.

[1] https://gitlab.com/yaook/k8s/-/blob/983cd725a3ecb29cdf735af3340e5c9f663ac819/k8s-core/ansible/vars/vault-config.yaml