flake: remove announcement of extra-substituters
Flakes may announce substituters (binary caches) which contain packages
used by that flake. When the flakes are used, Nix will then ask the user
whether to automatically add the substituters to the system or user wide
configuration. This is merely a convenience feature.

One problem with this is that we use the flake within a direnv hook, so
entering a directory can be stalled with a (y/N) question which may not
be directly obvious to the user.

Another problem is that Nix only allows auto configuring substituters
when the user has added themselves as a trusted user which most have
not.

This commit removes the announcement of extra-substituters. They can
still be manually configured as described in the documentation.
Substituters are always optional. If a package is not found in any
binary cache, it will be built from source.