direnv: honor MINIMAL_ACCESS_VENV for Nix layout
This commit adds a new minimal devShell to flake.nix that is
automatically used when MINIMAL_ACCESS_VENV is set to true.

Currently, there is no way to configure poetry2nix such that the default
dependencies are not installed (there is no equivalent to `poetry
install --only`), which means that our minimal-access venv does not work
as expected.

There is an upstream PR pending[1], but in order to not depend on a fork
until it is merged, all default dependencies are moved to a group called
`main` as a workaround. Actually, according to the poetry docs, the
"dependencies declared in tool.poetry.dependencies are part of an
implicit main group"[2], so everything else should work just like
before.

[1] https://github.com/nix-community/poetry2nix/pull/1785
[2]
https://python-poetry.org/docs/managing-dependencies#dependency-groups