Conditionally set Wireguard MTU on server and client
The server side is clever and automatically discovers a proper MTU if
the default will not work out. However, we explicitly must set the MTU
for clients then as otherwise connectivity is broken. To be able to
properly calculate the MTU, we must expose the network_mtu to Ansible.

The MTU can be calculated in the following way:
network_mtu - IPv4/6 header - UDP header - Wireguard header
network_mtu - 20/40 - 8 - 32

On the client side we limit the MTU to 1492 which is the maximum of a
usual DSL connection.