Drop C&H specific Terraform defaults
Some input variables used default values that are highly specific to the
underlying Openstack cloud, such as image and flavor names. It is not
useful to preset them.

Thus, Cloud&Heat specific default have been moved from the Terraform
default values to the template file where they act as both an example
and a way for company users to quickly set up a cluster.