Fix spawning clusters on Openstack with apply-all.sh
8290d8151234190ed77914e0d6741c465b7454f7 (!1340) moved the definition of
some variables behind a function to make sourcing lib.sh side-effect
free.

As that commit was originally part of (!1265) where the inventory is
updated early in apply-all.sh and the values are read from
config.tfvars.json instead of from config.toml, a condition has been
added that depends on the existance of config.tfvars.json to be present
for the tf_usage and terraform_prevent_disruption to be set. The
condition should have been removed when rebasing the commit to devel and
changing the source to config.toml.

This commit removes the condition, so that both variables are always set
according to config.toml

Closes #705