Remove workaround for openstack in CI
Due to openstack virtual machines sometimes not becoming available via
SSH in our CI clusters, we use a script in our CI pipeline that checks
for reachability and in case of failure tries to re-create all VMs.

This, however, resulted in a bug (#705) not being found in the CI
because apply-terraform.sh has to be run before apply-all.sh.

As we now re-try all cluster-test jobs, we may not need this workaround
anymore. Thus, this commit is an attempt to bring the CI more in line
with what is actually done when spawning a cluster manually by removing
the workaround script entirely.